requirements.txt
-i https://pypi.org/simple
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-celery-results==1.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-ipware==2.1.0
django-structlog==1.5.1
django-templated-mail==1.1.1
django==2.2.11
djangorestframework==3.11.0
djoser==2.0.3
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0 ; python_version < '3.8'
inflection==0.3.1
itypes==1.1.0
jinja2==2.11.1
kombu==4.6.8
markdown==3.2.1
markupsafe==1.1.1
packaging==20.3
plaid-python==3.7.0
psycopg2-binary==2.8.5
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml.clib==0.2.0 ; platform_python_implementation == 'CPython' and python_version < '3.9'
ruamel.yaml==0.16.10
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
structlog==20.1.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
